+++ source/win-main/sidebar/ToCTab.vue
@@ -6,11 +6,17 @@
     <div
       v-for="(entry, idx) of tableOfContents"
       v-bind:key="idx"
+      v-bind:data-line="entry.line"
       class="toc-entry-container"
+      draggable="true"
       v-bind:style="{
         'margin-left': `${entry.level * 10}px`
       }"
       v-on:click="($root as any).jtl(entry.line, true)"
+      v-on:dragstart="startDragging"
+      v-on:dragenter="dragEnter"
+      v-on:dragleave="dragLeave"
+      v-on:drop="drop"
     >
       <div class="toc-level">
         {{ entry.renderedLevel }}
@@ -30,14 +36,16 @@ import { MDFileMeta } from '@dts/common/fsal'
 import { defineComponent } from 'vue'
 import sanitizeHtml from 'sanitize-html'
 import { getConverter } from '@common/util/md-to-html'
+import { ToCEntry } from '@common/modules/markdown-editor/util/generate-toc'
 
 // Must be instantiated after loading, i.e. when the Sidebar is initialized
 let md2html: Function
 
 export default defineComponent({
   name: 'ToCTab',
+  emits: ['move-section'],
   computed: {
-    tableOfContents: function (): any|null {
+    tableOfContents: function (): ToCEntry[]|null {
       return this.$store.state.tableOfContents
     },
     /**
@@ -79,6 +87,10 @@ export default defineComponent({
      * @param   {number}  tocEntryIdx           Index of heading in ToC
      */
     tocEntryIsActive: function (tocEntryLine: number, tocEntryIdx: number) {
+      if (this.tableOfContents === null) {
+        return false
+      }
+
       const cursorLine = this.$store.state.activeDocumentInfo.cursor.line
 
       // Determine index of next heading in ToC list
@@ -106,8 +118,75 @@ export default defineComponent({
         // Headings may be emphasised and contain code
         allowedTags: [ 'em', 'kbd', 'code' ]
       })
+    },
+    startDragging: function (event: DragEvent) {
+      if (event.currentTarget === null) {
+        return
+      }
+      const fromLine = (event.currentTarget as HTMLElement).dataset.line
+      event.dataTransfer?.setData('x-zettlr/toc-drag', fromLine as string)
+    },
+    dragEnter: function (event: DragEvent) {
+      if (event.currentTarget === null) {
+        return
+      }
+      const container = event.currentTarget as HTMLElement
+      container.classList.add('toc-drop-effect')
+    },
+    dragLeave: function (event: DragEvent) {
+      if (event.currentTarget === null) {
+        return
+      }
+      const container = event.currentTarget as HTMLElement
+      container.classList.remove('toc-drop-effect')
+    },
+    drop: function (event: DragEvent) {
+      if (event.currentTarget === null) {
+        return
+      }
+
+      const container = event.currentTarget as HTMLElement
+      container.classList.remove('toc-drop-effect')
+
+      const fromLine = parseInt(event.dataTransfer?.getData('x-zettlr/toc-drag') as string, 10)
+      const toLine = parseInt(container.dataset.line as string, 10)
+      const actualToLine = this.findEndOfEntry(toLine)
+      if (actualToLine === undefined) {
+        console.warn('Could not move section: Could not find correct target line')
+        return
+      }
+
+      this.$emit('move-section', { from: fromLine, to: actualToLine })
+    },
+    findEndOfEntry: function (originalToLine: number) {
+      if (this.tableOfContents === null) {
+        return
+      }
+
+      const idx = this.tableOfContents.findIndex(elem => elem.line === originalToLine)
+
+      if (idx < 0) {
+        return
+      }
+
+      if (idx === this.tableOfContents.length - 1) {
+        return -1
+      } else {
+        return this.tableOfContents[idx + 1].line - 1
+      }
     }
   }
 })
 
 </script>
+
+<style lang="less">
+// Add a neat little effect to the table of content entries as you drag them
+.toc-entry-container {
+  border-bottom: 2px solid transparent;
+
+  &.toc-drop-effect {
+    border-bottom-color: rgb(40, 100, 255);
+  }
+}
+</style>